made amendments

+++ src/main/java/ru/Rakhmatov/Lesson3/AgeComparator.java
@@ -14,6 +14,6 @@ public class AgeComparator implements Comparator<Employee>{
     @Override
     public int compare(Employee o1, Employee o2) {
         return sortType == SortType.Ascending ? Integer.compare(o1.getAge(), o2.getAge()) :
-                Double.compare(o2.getAge(), o1.getAge());
+                Integer.compare(o2.getAge(), o1.getAge());
     }
 }
